# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "description": "API description",
        "contact": {
            "email": "contact@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/provider-user-roles": {
            "get": {
                "tags": [
                    "Provider User Roles"
                ],
                "summary": "list of provider user roles",
                "description": "Returns the entire list of provider user roles",
                "operationId": "ProviderUserRole.all",
                "responses": {
                    "200": {
                        "description": "Operation successful",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Provider User Roles"
                ],
                "summary": "Create a new provider user role",
                "description": "__*Security:*__ __*can be used only by clients with 'admin' role*__",
                "operationId": "ProviderUserRole.create",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "provider_id": {
                                        "description": "Provider id",
                                        "type": "integer"
                                    },
                                    "user_id": {
                                        "description": "User id",
                                        "type": "integer"
                                    },
                                    "role_id": {
                                        "description": "Role id",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Operation successful",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/v1/provider-user-roles/{id}": {
            "get": {
                "tags": [
                    "Provider User Roles"
                ],
                "summary": "Returns provider user role by id",
                "description": "Returns provider user role details by id",
                "operationId": "ProviderUserRole.find",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Provider user role id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation successful",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Provider User Roles"
                ],
                "summary": "Update provider user role by id",
                "description": "__*Security:*__ __*can be used only by clients with 'admin' role*__",
                "operationId": "ProviderUserRole.update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Provider user role id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "provider_id": {
                                        "description": "Provider id",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "user_id": {
                                        "description": "User id",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "role_id": {
                                        "description": "Role id",
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operation successful",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Provider User Roles"
                ],
                "summary": "Delete provider user role by id",
                "description": "__*Security:*__ __*can be used only by clients with 'admin' role*__",
                "operationId": "ProviderUserRole.delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Provider user role id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Operation successful",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "OAuth2 Client Credentials Flow",
                "flows": {
                    "clientCredentials": {
                        "tokenUrl": "/oauth/token",
                        "scopes": {}
                    }
                }
            }
        }
    },
...
}